pyproject.toml
[tool.poetry]
name = "material-science-expert"
version = "0.1.0"
description = "LLM wrapper for llama3.2:1b with context information from Material Science project"
authors = ["saj3sh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.15"
streamlit = "1.32"
langchain-ollama = "^0.2.0"
pydantic = "2.9.2"   # version 2.10 has breaking changes
mp-api = "^0.42.2"
qdrant-client = "^1.12.1"
langchain-community = "^0.3.7"
huggingface-hub = "^0.26.2"
transformers = "^4.46.3"
torch = "^2.5.1"
python-dotenv = "^1.0.1"
langchain-qdrant = "^0.2.0"
watchdog = "^6.0.0"
langgraph = "^0.2.53"
ipython = "^8.29.0"
duckduckgo-search = "5.3.1b1"
streamlit-aggrid = "^1.0.5"
markdown = "^3.7"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
html5lib = "^1.1"


[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
